requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
entrypoints==0.4
ffmpegio==0.8.3
ffmpegio-core==0.8.3
filelock==3.11.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.3.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pluggy==1.0.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pyarrow==11.0.0
pydeck==0.8.1b0
Pygments==2.15.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.4
six==1.16.0
smmap==5.0.0
streamlit==1.21.0
sympy==1.11.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
zipp==3.15.0
altair==4.2.2
attrs==22.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
entrypoints==0.4
ffmpegio==0.8.3
ffmpegio-core==0.8.3
filelock==3.11.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.3.0
Jinja2==3.1.2
jsonschema==4.17.3
kaleido==0.2.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.1
pluggy==1.0.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pyarrow==11.0.0
pydeck==0.8.1b0
Pygments==2.15.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.4
six==1.16.0
smmap==5.0.0
streamlit==1.21.0
sympy==1.11.1
tenacity==8.2.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
